requirements.txt
PILLOW
opencv-contrib-python
numpy